feat: move report module under backtesting module
