+++ src/strategy_tester/backtesting/report/html/_dataframe_to_candlestickchart_options.py
