+++ src/strategy_tester/backtesting/report/html/_dataframe_to_histogramchart_options.py
