+++ src/strategy_tester/backtesting/report/html/_dataframe_to_scatterchart_cartesian_options.py
