+++ src/strategy_tester/backtesting/report/html/_dataseries_to_linechart_options.py
