+++ src/strategy_tester/backtesting/report/html/_dataseries_to_stacked_barchart_options_splitted_by_zeroline.py
