+++ src/strategy_tester/backtesting/report/html/_heatmap.py
