+++ src/strategy_tester/backtesting/report/html/_statistics_to_json.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from ...utils import parse_data_for_json
+from ....utils import parse_data_for_json
 
 
 def _statistics_to_json(stats: pd.Series, pnl: pd.Series):
