+++ src/strategy_tester/backtesting/report/pdf/__init__.py
