+++ src/strategy_tester/backtesting/report/pdf/_add_equity.py
