+++ src/strategy_tester/backtesting/report/pdf/_add_statistics.py
