+++ src/strategy_tester/backtesting/report/pdf/report_pdf.py
@@ -1,6 +1,6 @@
 from matplotlib.backends.backend_pdf import PdfPages
 
-from ...pipeline.context import Context
+from ....pipeline.context import Context
 from ._add_equity import _add_equity
 from ._add_heatmap import _add_heatmap
 from ._add_metadata import _add_metadata
